Return Nonce from ID Token response
We need the nonce from the OpenID Connect ID token in order to validate
that the token sent to us is secure. We will verify that the nonce we
send with the intial auth request is the one returned to us and ensure
it can only be used once.

https://learn.microsoft.com/en-us/entra/identity-platform/id-tokens#validate-tokens

Issue: https://khanacademy.atlassian.net/browse/INFRA-10699